+++ packages/code-analyzer-regex-engine/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@salesforce/code-analyzer-regex-engine",
   "description": "Regex Engine for the Salesforce Code Analyzer",
-  "version": "0.2.0",
+  "version": "0.1.0",
   "author": "The Salesforce Code Analyzer Team",
   "license": "BSD-3-Clause license",
   "homepage": "https://developer.salesforce.com/docs/platform/salesforce-code-analyzer/overview",
@@ -13,7 +13,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "dependencies": {
-    "@salesforce/code-analyzer-engine-api": "0.1.0",
+    "@salesforce/code-analyzer-engine-api": "^0.1.0",
     "@types/node": "^20.0.0"
   },
   "devDependencies": {
